Alte MapInfo und Speicherroutinen werden wieder standardmäßig genutzt, damit das Programm lauffähig ist
Sonst ein paar kleine Änderungen... :-)
z.B. einzelne Monitore für jede Kante beim ArcFlags setzen
z.B. aktuellen Wert eines Timers bekommen (Util.getTimer)

+++ java/src/kit/ral/common/Coordinates.java
@@ -89,7 +89,7 @@ public class Coordinates {
      * @param factor
      * @return this
      */
-    public Coordinates scale(float factor) {
+    public Coordinates scale(double factor) {
         latitude *= factor;
         longitude *= factor;
         return this;
